CRC: move block_proposal_after_pushed_is_rejected to its own file
